chore: convert models.py into separate modules
