+++ bfportal/core/models/__init__.py
@@ -1,0 +1,5 @@
+# flake8: noqa F401
+from .categories import ExperiencesCategory, SubCategory
+from .experience import ExperiencePage, ExperiencesPage
+from .pages import BlogPage, CustomBasePage, HomePage
+from .users import Profile, ProfilePage
